dhcpv6: make allocation for ia_pd entries account for worst case
One could gather, sort and count unique entries, at a linear algorithmic cost,
but such an implementation is overall slower for the majority who receive a
small amount of prefixes.

If we are to account for the worst case prefix exclusion which can be up to 128
bits, then we should pre-allocate not 10 bytes, but 21. This is for the future,
because odhcp6c today handles exclusions of 32-64 bits.

Signed-off-by: Paul Donald <newtwen+github@gmail.com>